+++ Changelog.md
@@ -6,6 +6,7 @@
 - Fix: Redstone does not automatically connect to the Alarm, Projector, and Username Logger
 - Fix: Some text in the check password screens doesn't show up properly
 - Fix: Options are not synchronized between newly placed laser blocks
+- Fix: Modules are not properly synchronized between laser blocks in certain situations
 
 --------------------------Changelog for v1.9.5 of SecurityCraft--------------------------